feat: added key, scale, and bpm
